added modified file i forgot to save for previous commit

+++ JobApplicationTrackerWebsite/server/src/controllers/applicationController.ts
@@ -99,6 +99,14 @@ export async function deleteApplication(req: Request, res: Response) {
     try {
         const id = Number(req.params.id);
 
+        if(!Number.isInteger(id) || id <= 0){
+            return res.status(400).json({
+                message: "Invalid application ID."
+            });
+        }
+
+        
+
         await prisma.jobApplication.delete({
             where: {
                 id,
